add Type conversion for complex arithmetic expression

+++ engine/src/main/java/org/opencds/cqf/cql/engine/elm/execution/ToLongEvaluator.java
@@ -28,7 +28,7 @@ public class ToLongEvaluator extends org.cqframework.cql.elm.execution.ToLong {
         }
 
         if (operand instanceof Integer) {
-            return (Long)operand;
+            return Long.valueOf((Integer)operand);
         }
 
         if (operand instanceof String) {
